Rename ui session manager service
