+++ ui/main/src/app/modules/core/session-end/session-end.component.ts
@@ -10,7 +10,7 @@
 import {ChangeDetectionStrategy, ChangeDetectorRef, Component, OnInit} from '@angular/core';
 import {I18n} from '@ofModel/i18n.model';
 import {ModalService} from 'app/business/services/modal.service';
-import {SessionManagerService} from 'app/business/services/session-manager.service';
+import {SessionManagerService} from '@ofServices/sessionManager/SessionManagerService';
 import {NgIf} from '@angular/common';
 import {TranslateModule} from '@ngx-translate/core';
